+++ regression/reference/classes/wrapClass1.cpp
@@ -80,8 +80,8 @@ CLA_Class1 * CLA_Class1_ctor_flag(int flag, CLA_Class1 * SHC_rv)
 // start CLA_Class1_delete
 void CLA_Class1_delete(CLA_Class1 * self)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.delete
     delete SH_this;
     self->addr = nullptr;
@@ -100,8 +100,8 @@ void CLA_Class1_delete(CLA_Class1 * self)
 // start CLA_Class1_method1
 int CLA_Class1_method1(CLA_Class1 * self)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.method1
     int SHC_rv = SH_this->Method1();
     return SHC_rv;
@@ -146,8 +146,8 @@ bool CLA_Class1_equivalent(const CLA_Class1 * self, CLA_Class1 * obj2)
 // start CLA_Class1_return_this
 void CLA_Class1_return_this(CLA_Class1 * self)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.return_this
     SH_this->returnThis();
     // splicer end class.Class1.method.return_this
@@ -174,8 +174,8 @@ void CLA_Class1_return_this(CLA_Class1 * self)
 CLA_Class1 * CLA_Class1_return_this_buffer(CLA_Class1 * self,
     char * name, bool flag, CLA_Class1 * SHC_rv)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.return_this_buffer
     std::string SHCXX_name(name);
     classes::Class1 * SHCXX_rv = SH_this->returnThisBuffer(SHCXX_name,
@@ -207,8 +207,8 @@ CLA_Class1 * CLA_Class1_return_this_buffer(CLA_Class1 * self,
 CLA_Class1 * CLA_Class1_return_this_buffer_bufferify(CLA_Class1 * self,
     char * name, int Lname, bool flag, CLA_Class1 * SHC_rv)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.return_this_buffer_bufferify
     std::string SHCXX_name(name, Lname);
     classes::Class1 * SHCXX_rv = SH_this->returnThisBuffer(SHCXX_name,
@@ -254,8 +254,8 @@ CLA_Class1 * CLA_Class1_getclass3(const CLA_Class1 * self,
 // start CLA_Class1_get_name
 const char * CLA_Class1_get_name(CLA_Class1 * self)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.get_name
     const std::string & SHCXX_rv = SH_this->getName();
     const char * SHC_rv = SHCXX_rv.c_str();
@@ -280,8 +280,8 @@ const char * CLA_Class1_get_name(CLA_Class1 * self)
 void CLA_Class1_get_name_bufferify(CLA_Class1 * self,
     CLA_SHROUD_array *DSHF_rv)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.get_name_bufferify
     const std::string & SHCXX_rv = SH_this->getName();
     ShroudStrToArray(DSHF_rv, &SHCXX_rv, 0);
@@ -300,8 +300,8 @@ void CLA_Class1_get_name_bufferify(CLA_Class1 * self,
 // start CLA_Class1_direction_func
 int CLA_Class1_direction_func(CLA_Class1 * self, int arg)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.direction_func
     classes::Class1::DIRECTION SHCXX_arg =
         static_cast<classes::Class1::DIRECTION>(arg);
@@ -320,8 +320,8 @@ int CLA_Class1_direction_func(CLA_Class1 * self, int arg)
 // start CLA_Class1_get_m_flag
 int CLA_Class1_get_m_flag(CLA_Class1 * self)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.get_m_flag
     return SH_this->m_flag;
     // splicer end class.Class1.method.get_m_flag
@@ -335,8 +335,8 @@ int CLA_Class1_get_m_flag(CLA_Class1 * self)
 // start CLA_Class1_get_test
 int CLA_Class1_get_test(CLA_Class1 * self)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.get_test
     return SH_this->m_test;
     // splicer end class.Class1.method.get_test
@@ -354,8 +354,8 @@ int CLA_Class1_get_test(CLA_Class1 * self)
 // start CLA_Class1_set_test
 void CLA_Class1_set_test(CLA_Class1 * self, int val)
 {
-    classes::Class1 *SH_this =
-        static_cast<classes::Class1 *>(self->addr);
+    classes::Class1 *SH_this = static_cast<classes::Class1 *>
+        (self->addr);
     // splicer begin class.Class1.method.set_test
     SH_this->m_test = val;
     return;