+++ regression/reference/forward/wrapClass2.cpp
@@ -38,8 +38,8 @@ FOR_Class2 * FOR_Class2_ctor(FOR_Class2 * SHC_rv)
 // Exact:     c_shadow_dtor
 void FOR_Class2_dtor(FOR_Class2 * self)
 {
-    forward::Class2 *SH_this =
-        static_cast<forward::Class2 *>(self->addr);
+    forward::Class2 *SH_this = static_cast<forward::Class2 *>
+        (self->addr);
     // splicer begin class.Class2.method.dtor
     delete SH_this;
     self->addr = nullptr;
@@ -56,8 +56,8 @@ void FOR_Class2_dtor(FOR_Class2 * self)
 // Match:     c_shadow_in
 void FOR_Class2_func1(FOR_Class2 * self, TUT_Class1 * arg)
 {
-    forward::Class2 *SH_this =
-        static_cast<forward::Class2 *>(self->addr);
+    forward::Class2 *SH_this = static_cast<forward::Class2 *>
+        (self->addr);
     // splicer begin class.Class2.method.func1
     tutorial::Class1 * SHCXX_arg = static_cast<tutorial::Class1 *>
         (arg->addr);
@@ -75,8 +75,8 @@ void FOR_Class2_func1(FOR_Class2 * self, TUT_Class1 * arg)
 // Match:     c_shadow_in
 void FOR_Class2_accept_class3(FOR_Class2 * self, FOR_Class3 * arg)
 {
-    forward::Class2 *SH_this =
-        static_cast<forward::Class2 *>(self->addr);
+    forward::Class2 *SH_this = static_cast<forward::Class2 *>
+        (self->addr);
     // splicer begin class.Class2.method.accept_class3
     forward::Class3 * SHCXX_arg = static_cast<forward::Class3 *>
         (arg->addr);