+++ regression/run/struct/struct.h
@@ -82,6 +82,8 @@ struct Cstruct_as_numpy {
 };
 typedef struct Cstruct_as_numpy Cstruct_as_numpy;
 
+Cstruct_as_class *Create_Cstruct_as_class(void);
+
 int acceptBothStructs(Cstruct_as_class *s1, Cstruct_as_numpy *s2);
 
 Cstruct_list *get_global_struct_list(void);